Move source under pkg dir; Rework dockerfile
