+++ main.go
@@ -8,16 +8,16 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/redhatinsights/insights-ingress-go/announcers"
-	"github.com/redhatinsights/insights-ingress-go/config"
-	l "github.com/redhatinsights/insights-ingress-go/logger"
-	"github.com/redhatinsights/insights-ingress-go/queue"
-	"github.com/redhatinsights/insights-ingress-go/stage"
-	"github.com/redhatinsights/insights-ingress-go/stage/minio"
-	"github.com/redhatinsights/insights-ingress-go/stage/s3"
-	"github.com/redhatinsights/insights-ingress-go/upload"
-	"github.com/redhatinsights/insights-ingress-go/validators/kafka"
-	"github.com/redhatinsights/insights-ingress-go/version"
+	"github.com/redhatinsights/insights-ingress-go/pkg/announcers"
+	"github.com/redhatinsights/insights-ingress-go/pkg/config"
+	l "github.com/redhatinsights/insights-ingress-go/pkg/logger"
+	"github.com/redhatinsights/insights-ingress-go/pkg/queue"
+	"github.com/redhatinsights/insights-ingress-go/pkg/stage"
+	"github.com/redhatinsights/insights-ingress-go/pkg/stage/minio"
+	"github.com/redhatinsights/insights-ingress-go/pkg/stage/s3"
+	"github.com/redhatinsights/insights-ingress-go/pkg/upload"
+	"github.com/redhatinsights/insights-ingress-go/pkg/validators/kafka"
+	"github.com/redhatinsights/insights-ingress-go/pkg/version"
 
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
