+++ pkg/announcers/fake.go
@@ -3,7 +3,7 @@ package announcers
 import (
 	"sync"
 
-	l "github.com/redhatinsights/insights-ingress-go/logger"
+	l "github.com/redhatinsights/insights-ingress-go/pkg/logger"
 	"github.com/sirupsen/logrus"
 )
