+++ pkg/stage/minio/types.go
