+++ pkg/upload/upload_suite_test.go
@@ -5,7 +5,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	l "github.com/redhatinsights/insights-ingress-go/logger"
+	l "github.com/redhatinsights/insights-ingress-go/pkg/logger"
 )
 
 func TestUpload(t *testing.T) {
