+++ pkg/version/version_suite_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	l "github.com/redhatinsights/insights-ingress-go/logger"
+	l "github.com/redhatinsights/insights-ingress-go/pkg/logger"
 )
 
 func TestInventory(t *testing.T) {
